roachtest: add placeholders hints for create_tenant
Fixes: #91903
Fixes: #91905

Release note: None

+++ pkg/cmd/roachtest/tests/multitenant_tpch.go
@@ -72,7 +72,7 @@ func runMultiTenantTPCH(ctx context.Context, t test.Test, c cluster.Cluster) {
 		tenantSQLPort  = 30258
 		tenantNode     = 1
 	)
-	_, err := singleTenantConn.Exec(`SELECT crdb_internal.create_tenant($1)`, tenantID)
+	_, err := singleTenantConn.Exec(`SELECT crdb_internal.create_tenant($1::INT)`, tenantID)
 	if err != nil {
 		t.Fatal(err)
 	}
